chore: implement latest NodeSecure & start cleanup
