+++ package.json
@@ -3,6 +3,7 @@
     "version": "1.0.0",
     "description": "NodeSecure HTML & PDF graphic security report",
     "main": "index.js",
+    "type": "module",
     "bin": {
         "nreport": "./bin/index.js"
     },
@@ -25,7 +26,6 @@
         "chart"
     ],
     "author": "NodeSecure",
-    "type": "module",
     "license": "MIT",
     "bugs": {
         "url": "https://github.com/NodeSecure/report/issues"
@@ -41,6 +41,7 @@
         "dotenv": "^16.0.0",
         "esbuild": "^0.14.27",
         "filenamify": "^5.1.0",
+        "frequency-set": "^1.0.1",
         "isomorphic-git": "^1.14.0",
         "kleur": "^4.1.3",
         "puppeteer": "^13.5.1",
